+++ src/varintExternal.c
@@ -105,7 +105,7 @@ static void varintExternalCopyToEncodingBigEndian_(uint8_t *restrict dst,
     /* so, for encoding of 8 bytes, this does:
      *  dst[0] = src[7]; dst[1] = src[6]; ...; dst[7] = src[0] */
     uint8_t resPos = 0;
-    uint8_t srcPos = encoding;
+    uint8_t srcPos = (uint8_t)encoding;
 
     while (srcPos > 0) {
         dst[resPos++] = src[--srcPos];
@@ -147,7 +147,7 @@ varintExternalLoadFromEncodingBigEndian_(const uint8_t *restrict src,
      * we stop when we run out of encoded bytes
      * (number of bytes == encoding value) */
     uint8_t resPos = 0;
-    uint8_t srcPos = encoding;
+    uint8_t srcPos = (uint8_t)encoding;
 
     while (srcPos > 0) {
         resarr[resPos++] = src[--srcPos];
@@ -180,7 +180,7 @@ varintBigExternalLoadFromEncodingBigEndian_(const uint8_t *restrict src,
      * we stop when we run out of encoded bytes
      * (number of bytes == encoding value) */
     uint8_t resPos = 0;
-    uint8_t srcPos = encoding;
+    uint8_t srcPos = (uint8_t)encoding;
 
     while (srcPos > 0) {
         resarr[resPos++] = src[--srcPos];
@@ -199,7 +199,7 @@ varintWidth varintExternalSignedEncoding(int64_t value) {
     }
 
     varintWidth encoding;
-    varintExternalUnsignedEncoding(value, encoding);
+    varintExternalUnsignedEncoding((uint64_t)value, encoding);
     return encoding;
 }
 
@@ -260,7 +260,7 @@ static varintWidth varintExternalAdd_(uint8_t *p, varintWidth origEncoding,
     VARINT_ADD_OR_ABORT_OVERFLOW_(updatingVal, add, newVal);
 
     varintWidth newEncoding;
-    varintExternalUnsignedEncoding(updatingVal, newEncoding);
+    varintExternalUnsignedEncoding((uint64_t)updatingVal, newEncoding);
 
     /* If new encoding is larger than current encoding, we don't
      * want to overwrite memory beyond our current varint.
@@ -269,7 +269,7 @@ static varintWidth varintExternalAdd_(uint8_t *p, varintWidth origEncoding,
         return newEncoding;
     }
 
-    varintExternalPut(p, newVal);
+    varintExternalPut(p, (uint64_t)newVal);
     return newEncoding;
 }
